Move ServiceDiscovery plugin interface out of internal/ path #38
